feat(kit,nuxi): add `writeTypes` utility (#22385)
