+++ packages/kit/src/internal/cjs.ts
@@ -60,15 +60,19 @@ function getRequireCacheItem (id: string) {
   }
 }
 
+export function getModulePaths (paths?: string[] | string) {
+  return ([] as Array<string | undefined>).concat(
+    global.__NUXT_PREPATHS__,
+    paths || [],
+    process.cwd(),
+    global.__NUXT_PATHS__
+  ).filter(Boolean) as string[]
+}
+
 /** @deprecated Do not use CJS utils */
 export function resolveModule (id: string, opts: ResolveModuleOptions = {}) {
   return normalize(_require.resolve(id, {
-    paths: ([] as Array<string | undefined>).concat(
-      global.__NUXT_PREPATHS__,
-      opts.paths || [],
-      process.cwd(),
-      global.__NUXT_PATHS__
-    ).filter(Boolean) as string[]
+    paths: getModulePaths(opts.paths)
   }))
 }
